+++ CHANGELOG.md
@@ -22,6 +22,7 @@ Caso adicionado no escopo inicial, todos os conteúdos abaixo não serão adicio
 
 ### Corrigido
 - `QasActions`: Corrigido tamanho do botão quando passado via props e utilizado junto com `useEqualWidth`.
+- `QasNumericInput`: Corrigido exibição do input, no qual exibia o conteúdo da lib `Autonumeric` ao dar hover, sendo que deve ser exibida apenas quando tem foco ou valor no campo.
 - `QasSelectListDialog`:
   - Corrigido slots do dialog que não funcionavam.
   - Corrigido alinhamento do botão de adicionar item quando não se tem label no header.